Code refactoring

+++ RSS-Feed-Reader/src/main/java/de/htwsaar/pib2021/rss_feed_reader/commands/FeedItemCommand.java
@@ -14,6 +14,6 @@ public class FeedItemCommand {
     private String channelCategory;
     private Integer elapsedPublishMinutes;
     private Integer elapsedPublishHoures;
-    private Integer elapsedPublishDays;
+    private Long elapsedPublishDays;
 
 }
